Separate out `admin.logs.v1` feature
