+++ apps/console/package.json
@@ -67,6 +67,7 @@
         "@wso2is/admin.extensions.v1": "^2.22.12",
         "@wso2is/admin.groups.v1": "^2.20.124",
         "@wso2is/admin.identity-verification-providers.v1": "^2.20.124",
+        "@wso2is/admin.logs.v1": "^0.0.0",
         "@wso2is/admin.oidc-scopes.v1": "^2.20.124",
         "@wso2is/admin.org-insights.v1": "^2.20.124",
         "@wso2is/admin.organization-discovery.v1": "^2.20.124",
