+++ features/admin.logs.v1/package.json
@@ -1,0 +1,76 @@
+{
+    "private": true,
+    "name": "@wso2is/admin.logs.v1",
+    "version": "0.0.0",
+    "description": "Query your logs to troubleshoot issues and monitor resource activities.",
+    "author": "WSO2",
+    "license": "Apache-2.0",
+    "homepage": "https://github.com/wso2/identity-apps/tree/master/features/admin.logs.v1#readme",
+    "bugs": {
+        "url": "https://github.com/wso2/identity-apps/issues"
+    },
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/wso2/identity-apps",
+        "directory": "features/admin.logs.v1"
+    },
+    "keywords": [
+        "wso2",
+        "identity-server",
+        "features",
+        "admin.logs",
+        "v1"
+    ],
+    "dependencies": {
+        "@asgardeo/auth-react": "^5.0.3",
+        "@emotion/react": "^11.11.0",
+        "@emotion/styled": "^11.11.0",
+        "@mui/icons-material": "^5.11.16",
+        "@mui/lab": "5.0.0-alpha.129",
+        "@mui/material": "^5.13.0",
+        "@mui/system": "^5.12.3",
+        "@mui/utils": "^5.12.3",
+        "@oxygen-ui/react": "^1.13.6",
+        "@wso2is/admin.core.v1": "^2.24.15",
+        "@wso2is/core": "^2.0.58",
+        "@wso2is/form": "^2.1.2",
+        "@wso2is/forms": "^2.0.50",
+        "@wso2is/i18n": "^2.5.20",
+        "@wso2is/react-components": "^2.4.1",
+        "axios": "^0.19.2",
+        "i18next": "^21.9.1",
+        "moment": "^2.24.0",
+        "react-helmet": "^5.2.1",
+        "react-i18next": "^11.18.5",
+        "react-redux": "^7.2.9",
+        "redux": "^4.0.4",
+        "semantic-ui-react": "^2.1.3"
+    },
+    "devDependencies": {
+        "@rollup/plugin-commonjs": "^25.0.7",
+        "@rollup/plugin-dynamic-import-vars": "^2.1.2",
+        "@rollup/plugin-image": "^3.0.3",
+        "@rollup/plugin-json": "^6.1.0",
+        "@rollup/plugin-node-resolve": "^15.2.3",
+        "@rollup/plugin-typescript": "^11.1.6",
+        "@svgr/rollup": "^6.2.1",
+        "@types/react": "^18.0.18",
+        "@types/react-redux": "^7.1.25",
+        "rollup": "^4.17.2",
+        "rollup-plugin-dts": "^6.1.1",
+        "rollup-plugin-generate-package-json": "^3.2.0",
+        "rollup-plugin-polyfill-node": "^0.13.0",
+        "rollup-plugin-scss": "^4.0.0",
+        "rollup-plugin-styles": "^4.0.0",
+        "rollup-plugin-svg": "^2.0.0",
+        "typescript": "^4.6.4"
+    },
+    "peerDependencies": {
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-router-dom": "^4.3.1"
+    },
+    "browserslist": [
+        "> 0.2%"
+    ]
+}
