+++ features/admin.logs.v1/pages/audit-logs-page.tsx
@@ -15,14 +15,14 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
 import { getEmptyPlaceholderIllustrations } from "@wso2is/admin.core.v1";
 import { IdentifiableComponentInterface } from "@wso2is/core/models";
 import { EmptyPlaceholder,
     Hint,
     LinkButton,
     Message,
-    PrimaryButton } from "@wso2is/react-components";
+    PrimaryButton
+} from "@wso2is/react-components";
 import React, {
     MutableRefObject,
     ReactElement,
@@ -33,17 +33,17 @@ import React, {
 } from "react";
 import { useTranslation } from "react-i18next";
 import { Icon, Label } from "semantic-ui-react";
-import InfiniteScrollContainer from "../logs/components/infinite-scroll-container";
-import TimeRangeSelector from "../logs/components/time-range-selector";
-import { LogsConstants } from "../logs/constants";
-import useFetch from "../logs/hooks/use-fetch";
-import { InterfaceLogEntry, InterfaceLogsFilter, InterfaceLogsRequest, TabIndex } from "../logs/models/log-models";
-import { getDateTimeWithOffset, getTimeFromTimestamp } from "../logs/utils";
+import InfiniteScrollContainer from "../components/infinite-scroll-container";
+import TimeRangeSelector from "../components/time-range-selector";
+import { LogsConstants } from "../constants/logs-constants";
+import useFetch from "../hooks/use-fetch";
+import { InterfaceLogEntry, InterfaceLogsFilter, InterfaceLogsRequest, TabIndex } from "../models/log-models";
+import { getDateTimeWithOffset, getTimeFromTimestamp } from "../utils/datetime-utils";
 
 /**
  * Proptypes for the logs page component.
  */
-type DiagnosticPagePropsInterface = IdentifiableComponentInterface;
+type AuditPagePropsInterface = IdentifiableComponentInterface;
 
 /**
  * Logs monitoring page.
@@ -52,7 +52,7 @@ type DiagnosticPagePropsInterface = IdentifiableComponentInterface;
  *
  * @returns Logs Page {@link React.ReactElement}
  */
-export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactElement => {
+const AuditLogsPage = (props: AuditPagePropsInterface) : ReactElement => {
 
     const {
         ["data-componentid"]: componentId
@@ -71,11 +71,11 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
     const [ requestPayload, setRequestPayload ] = useState<InterfaceLogsRequest | null>(null);
     const [ showLastLogTime, setShowLastLogTime ] = useState<boolean>(false);
     const [ showDelayMessage, setShowDelayMessage ] = useState<boolean>(false);
-    const [ lastDiagnosticLogRequestTime, setLastDiagnosticLogRequestTime ] = useState<string>("");
-    const [ lastDiagnosticLogRequestTimeRange, setLastDiagnosticLogRequestTimeRange ] = useState<number>(0.25);
+    const [ lastAuditLogRequestTime, setLastAuditLogRequestTime ] = useState<string>("");
+    const [ lastAuditLogRequestTimeRange, setLastAuditLogRequestTimeRange ] = useState<number>(0.25);
     const [ showRefreshButton, setShowRefreshButton ] = useState<boolean>(false);
     const [ timerRunning, setTimerRunning ] = useState<boolean>(false);
-    const [ diagnosticLogList, setDiagnosticLogList ] = useState<InterfaceLogEntry[]>([]);
+    const [ audiLogList, setAuditLogList ] = useState<InterfaceLogEntry[]>([]);
 
     const { t } = useTranslation();
     const timeZone: string = "GMT+0000 UTC";
@@ -100,7 +100,7 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
                 }, LogsConstants.REFRESH_BUTTON_TIMEOUT);
             }
         }
-    }, [ lastDiagnosticLogRequestTime ]);
+    }, [ lastAuditLogRequestTime ]);
 
     useEffect(() => {
         setShowRefreshButton(false);
@@ -113,46 +113,38 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
 
         setEndTime(currentEndTime);
         setStartTime(currentStartTime);
-        setLastDiagnosticLogRequestTime(currentEndTime);
+        setLastAuditLogRequestTime(currentEndTime);
 
         // Fetch logs automatically during the first render.
         setRequestPayload({
             endTime: currentEndTime,
             filter: "",
             limit: LogsConstants.LOG_FETCH_COUNT,
-            logType: TabIndex.DIAGNOSTIC_LOGS,
+            logType: TabIndex.AUDIT_LOGS,
             startTime: currentStartTime
         });
     }, []);
 
     const scrollRef: MutableRefObject<HTMLDivElement> = useRef<HTMLDivElement>(null);
-
     const { error, list, loading, next, previous } = useFetch(requestPayload);
 
-
     useEffect (() => {
-        if (!loading && list.length > 0) {
-            setDiagnosticLogList(list);
-        }
-    }, [ list, loading ]);
+        setAuditLogList(list);
+    }, [ list ]);
 
     /**
-     * Handles the behaviour of the infinite scroller
+     * Handles the behavior of the infinite scroller.
      */
     const handleScroll: UIEventHandler<HTMLDivElement> = (e: React.UIEvent<HTMLElement>) => {
-        const element: any = e.target;
+        const element: HTMLElement = e.target as HTMLElement;
 
         setShowDelayMessage(false);
-        /**
-         * When the at the top of the log container
-         */
+        // When the at the top of the log container.
         if (element.scrollTop === 0) {
-
             if (previous) {
                 setRequestPayload({
                     filter: searchQuery,
                     limit: LogsConstants.LOG_FETCH_COUNT,
-                    logType: TabIndex.DIAGNOSTIC_LOGS,
                     previousToken: previous
                 });
                 setIsPreviousEmpty(false);
@@ -161,15 +153,12 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
                 setIsPreviousEmpty(true);
             }
         }
-        /**
-         * When at the bottom of the log container
-         */
+        // When the at the bottom of the log container.
         if (element.scrollHeight - element.scrollTop === element.clientHeight) {
             if (next) {
                 setRequestPayload({
                     filter: searchQuery,
                     limit: LogsConstants.LOG_FETCH_COUNT,
-                    logType: TabIndex.DIAGNOSTIC_LOGS,
                     nextToken: next
                 });
                 setIsNextEmpty(false);
@@ -180,7 +169,7 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
         }
     };
 
-    const renderDiagnosticLogContent = () : ReactElement => {
+    const renderAuditLogContent = () : ReactElement => {
 
         return (
             <div>
@@ -190,7 +179,7 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
                         setFromTime={ (value: string): void => setStartTime(value) }
                         setToTime={ (value: string): void => setEndTime(value) }
                         setTimeRange={ (value: number): void => setTimeRange(value) }
-                        data-componentid={ componentId }
+                        data-componentid={ `${ componentId }-audit` }
                     />
                     { showRefreshButton
                         ? (
@@ -213,30 +202,27 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
                     }
                 </div>
                 <div>
-                    <>
-                        <div className="top-toolbar">
-                            { renderRefreshTime() }
-                            <Label.Group>
-                                { filterList && filterList.map(
-                                    (value: { key: string, value:string }, index: number) =>
-                                        (<Label key={ index } className="filter-pill">
-                                            { getLabelTextForFilterPill(value.key) }
-                                            <Label.Detail>{ value.value }</Label.Detail>
-                                            <Icon name="delete" onClick={ () => removeFilter(value.key) }></Icon>
-                                        </Label>)
-                                ) }
-                                { resolveClearAllFilters() }
-                            </Label.Group>
-                        </div>
-                        { resolveDiagnosticLogs() }
-                    </>
+                    <div className="top-toolbar">
+                        { renderRefreshTime() }
+                        <Label.Group>
+                            { filterList?.map(
+                                (value: { key: string, value:string }, index: number) =>
+                                    (<Label key={ index } className="filter-pill">
+                                        { getLabelTextForFilterPill(value.key) }
+                                        <Label.Detail>{ value.value }</Label.Detail>
+                                        <Icon name="delete" onClick={ () => removeFilter(value.key) }></Icon>
+                                    </Label>)
+                            ) }
+                            { resolveClearAllFilters() }
+                        </Label.Group>
+                    </div>
+                    { resolveAuditLogs() }
                 </div>
-            </div>
-        );
+            </div> );
     };
 
     /**
-     * Build filter query
+     * Build filter query.
      */
     const buildFilterQuery = () => {
         let tempQuery: string = "";
@@ -252,9 +238,9 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
     };
 
     /**
-     * Search logs
+     * Search logs.
      *
-     * @param query - search query with filters
+     * @param query - search query with filters.
      */
     const handleSearch = () => {
         let currentQuery: string = "";
@@ -280,14 +266,14 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
             currentStartTime = (currentTime - 3600*1000*timeRange).toString();
         }
         setShowDelayMessage(false);
-        setLastDiagnosticLogRequestTime(currentTime.toString());
-        setLastDiagnosticLogRequestTimeRange(timeRange);
+        setLastAuditLogRequestTime(currentTime.toString());
+        setLastAuditLogRequestTimeRange(timeRange);
 
         setRequestPayload({
             endTime: currentEndTime,
             filter: `${filterQuery}${currentQuery}`,
             limit: LogsConstants.LOG_FETCH_COUNT,
-            logType: TabIndex.DIAGNOSTIC_LOGS,
+            logType: TabIndex.AUDIT_LOGS,
             startTime: currentStartTime
         });
     };
@@ -299,11 +285,10 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
         setSearchQuery("");
         setInputQuery("");
         setShowRefreshButton(false);
-
     };
 
     /**
-     * Handles searching logs with multiple filters
+     * Handles searching logs with multiple filters.
      */
     const handleFilter = (query: InterfaceLogsFilter): void => {
         const tempMap: Map<string, string> = filterMap;
@@ -323,7 +308,7 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
 
     /**
      * Remove specific filter attributes on click delete button
-     * in filter pills
+     * in filter pills.
      */
     const removeFilter = (filterKey: string): void => {
         const tempMap: Map<string, string> = filterMap;
@@ -342,7 +327,7 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
     };
 
     /**
-     * Clear all filter attributes
+     * Clear all filter attributes.
      */
     const clearAllFilters = (): void => {
         const tempMap: Map<string, string> = filterMap;
@@ -354,7 +339,7 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
     };
 
     /**
-     * Returns search component
+     * Returns search component.
      */
     const advancedSearchFilter = (): ReactElement => (
         <form
@@ -364,7 +349,7 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
                 <div className="search-box ui action left icon input advanced-search with-add-on">
                     <input
                         autoComplete="off"
-                        placeholder={ t("extensions:develop.monitor.filter.searchBar.placeholderDiagnostic") }
+                        placeholder={ t("extensions:develop.monitor.filter.searchBar.placeholderAudit") }
                         maxLength={ 120 }
                         name="query"
                         type="text"
@@ -395,7 +380,7 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
         </form>
     );
 
-    const resolveDiagnosticLogs = (): ReactElement => {
+    const resolveAuditLogs = (): ReactElement => {
         if (error) {
             if (error?.response?.data?.code === LogsConstants.END_TIME_GREATER_THAN_START_TIME_ERROR_CODE) {
                 return (
@@ -457,7 +442,7 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
                     data-componentid={ `${ componentId }-empty-search-placeholder` }
                 />
             );
-        } else if (filterQuery && list.length === 0) {
+        } else if (filterQuery && list?.length === 0) {
             return (
                 <EmptyPlaceholder
                     action={ (
@@ -475,7 +460,7 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
                     data-componentid={ `${ componentId }-empty-filter-placeholder` }
                 />
             );
-        } else if (list.length === 0 && (!next || !previous) && !loading) {
+        } else if (list?.length === 0 && (!next || !previous) && !loading) {
             return (
                 <EmptyPlaceholder
                     image={ getEmptyPlaceholderIllustrations().emptySearch }
@@ -485,7 +470,7 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
                         t("extensions:develop.monitor.notifications.emptyResponse.subtitle.0") +
                         `${ LogsConstants.TIMERANGE_DROPDOWN_OPTIONS.find(
                             (el: {key: number; text: string; value: number;}) =>
-                                el.value === lastDiagnosticLogRequestTimeRange)?.text.toLowerCase()
+                                el.value === lastAuditLogRequestTimeRange)?.text.toLowerCase()
                         }`,
                         t("extensions:develop.monitor.notifications.emptyResponse.subtitle.1")
                     ] }
@@ -496,7 +481,7 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
             return (
                 <>
                     { (loading && showDelayMessage) && (
-                        <div className="diagnostic-logs-delay-message">
+                        <div className="audit-logs-delay-message">
                             <Message
                                 type="info"
                                 compact
@@ -507,13 +492,13 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
                     <InfiniteScrollContainer
                         handleScroll={ handleScroll }
                         scrollRef={ scrollRef }
-                        logs={ diagnosticLogList }
+                        logs={ audiLogList }
                         loading={ loading }
                         rowHeight= { LogsConstants.LOG_ROW_HEIGHT }
                         logCount={ LogsConstants.LOG_FETCH_COUNT }
                         isPreviousEmpty = { isPreviousEmpty }
                         isNextEmpty = { isNextEmpty }
-                        logType={ TabIndex.DIAGNOSTIC_LOGS }
+                        logType={ TabIndex.AUDIT_LOGS }
                         setSearchQuery = { handleFilter }
                         data-componentid={ componentId }
                     />
@@ -543,7 +528,6 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
     };
 
     const fetchLatestLogs = (): void => {
-
         if (timeRange !== -1) {
             const current: number = getDateTimeWithOffset(timeZone);
 
@@ -554,7 +538,7 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
     };
 
     const renderRefreshTime = (): ReactElement => {
-        const isoDateString: string = new Date(Number(lastDiagnosticLogRequestTime)).toISOString();
+        const isoDateString: string = new Date(Number(lastAuditLogRequestTime)).toISOString();
 
         return (
             <div className="logs-refresh-component">
@@ -591,7 +575,17 @@ export const DiagnosticLogsPage = (props: DiagnosticPagePropsInterface) : ReactE
 
     return (
         <div>
-            { renderDiagnosticLogContent() }
+            { renderAuditLogContent() }
         </div>
     );
 };
+
+
+/**
+ * Default props for the component.
+ */
+AuditLogsPage.defaultProps = {
+    "data-componentid": "audit-logs"
+};
+
+export default AuditLogsPage;
