+++ features/admin.logs.v1/pages/logs-page.tsx
@@ -35,8 +35,8 @@ import React, {
 import { useTranslation } from "react-i18next";
 import { useSelector } from "react-redux";
 import { MenuItem, TabProps } from "semantic-ui-react";
-import { AuditLogsPage } from "../../audit-logs/audit-logs";
-import { DiagnosticLogsPage } from "../../diagnostic-logs/diagnostic-logs";
+import AuditLogsPage from "./audit-logs-page";
+import DiagnosticLogsPage from "./diagnostic-logs-page";
 import { TabIndex } from "../models/log-models";
 import "./logs.scss";
